Use `$PIPX_BIN_DIR` in `.zshrc`
The Zsh configuration file `.zshrc` in this repo adds some directories
to `$PATH`, including a directory where `pipx` binaries are installed.
`pipx` supports a `$PIPX_BIN_DIR` environment variable for such a task.
The environment variable value used here is the `pipx` default, but it
doesn't hurt to set the variable anyway to ensure that the exact same
directory is always on `$PATH`. This commit will set `$PIPX_BIN_DIR`.

https://pypa.github.io/pipx/installation/#installation-options